geminiProject/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
algoliasearch
algoliasearch-django
annotated-types
asgiref
attrs
cachetools
certifi
charset-normalizer
colorama
diff-match-patch
Django
django-cors-headers
django-haystack
django-import-export
djangorestframework
djangorestframework-simplejwt
drf-yasg
frozenlist
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
gunicorn
httplib2
idna
inflection
jsonpatch
jsonpointer
langchain
langchain-core
langchain-text-splitters
langsmith
multidict
mysqlclient
numpy
orjson
packaging
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
PyJWT
pyparsing
pytz
PyYAML
requests
rsa
setuptools
SQLAlchemy
sqlparse
tablib
tenacity
tqdm
typing_extensions
tzdata
uritemplate
urllib3
whitenoise
Whoosh
yarl
